# Turn optimizer_dpe_stats to on