--- a/src/test/regress/expected/qp_targeted_dispatch_optimizer.out
+++ b/src/test/regress/expected/qp_targeted_dispatch_optimizer.out
@@ -618,7 +618,6 @@ INFO:  (slice 1) Dispatch command to SINGLE content
 
 select count(*) from range_table where id=1;
 INFO:  (slice 1) Dispatch command to ALL contents: 0 1 2
-INFO:  (slice 2) Dispatch command to ALL contents: 0 1 2
  count 
 -------
      1